+++ packages/components/src/Form/Inputs/Select/SelectOptions.tsx
@@ -37,8 +37,8 @@ import styled from 'styled-components'
 import { Icon, IconPlaceholder } from '../../../Icon'
 import { Spinner } from '../../../Spinner'
 import { ListDivider } from '../../../List/ListDivider'
-import { ListItemDetail } from '../../../List/ListItemDetail'
-import { ListItemPreface } from '../../../List/ListItemPreface'
+import { ListItemDetail } from '../../../ListItem/ListItemDetail'
+import { ListItemPreface } from '../../../ListItem/ListItemPreface'
 import { Heading, HeadingProps, Paragraph, Text } from '../../../Text'
 import { useID } from '../../../utils'
 import {
