+++ packages/components/src/List/index.ts
@@ -26,10 +26,3 @@
 
 export * from './List'
 export * from './ListDivider'
-export * from './ListItem'
-export * from './ListItemContext'
-export * from './ListItemContent'
-export * from './ListItemIcon'
-export * from './ListItemLabel'
-export { listItemDimensions } from './utils'
-export type { ListItemProps } from './types'
