+++ packages/components/src/ListItem/ListItem.test.tsx
@@ -29,7 +29,7 @@ import React from 'react'
 import { renderWithTheme } from '@looker/components-test-utils'
 import { fireEvent, configure, screen } from '@testing-library/react'
 import { Science } from '@styled-icons/material-outlined/Science'
-import { List } from './List'
+import { List } from '../List'
 import { ListItem } from './ListItem'
 
 describe('ListItem', () => {
