+++ packages/components/src/ListItem/ListItemDetail.tsx
