+++ packages/components/src/ListItem/utils/createListItemPartitions.tsx
