+++ packages/components/src/ListItem/utils/getDetailOptions.ts
@@ -24,38 +24,20 @@
 
  */
 
-import { itemSelectedColor, Theme } from '@looker/design-tokens'
-import { css } from 'styled-components'
-import { FlexibleColor, ListItemStatefulProps } from '../types'
+import { Detail } from '../../ListItem/types'
 
-export const listItemBackgroundColor = ({
-  color,
-  disabled,
-  hovered,
-  selected,
-  theme: { colors },
-}: ListItemStatefulProps & FlexibleColor & { theme: Theme }) => {
-  const stateColors = color
-    ? {
-        all: colors[`${color}Subtle`],
-        hovered: colors.ui1,
-        selected: colors[`${color}Subtle`],
-      }
-    : {
-        all: itemSelectedColor(colors.ui2),
-        hovered: colors.ui1,
-        selected: itemSelectedColor(colors.ui2),
-      }
+// Simplifies the type check when dealing with ListItem and related components
+export const getDetailOptions = (detail: Detail) => {
+  let accessory, hoverDisclosure, content, width
 
-  let renderedColor
+  if (typeof detail === 'object' && detail && 'options' in detail) {
+    accessory = detail.options.accessory
+    content = detail.content
+    hoverDisclosure = detail.options.hoverDisclosure
+    width = detail.options.width
+  } else {
+    content = detail
+  }
 
-  if (disabled) renderedColor = 'transparent'
-  else if (selected && hovered) renderedColor = stateColors.all
-  else if (selected) renderedColor = stateColors.selected
-  else if (hovered) renderedColor = stateColors.hovered
-  else renderedColor = 'transparent'
-
-  return css`
-    background: ${renderedColor};
-  `
+  return { accessory, content, hoverDisclosure, width }
 }
