+++ packages/components/src/ListItem/utils/index.ts
@@ -24,33 +24,9 @@
 
  */
 
-import {
-  ListColor,
-  listItemColorAppliesToLabel,
-  listItemColorOptions,
-} from '../types'
-
-const listItemColor = (
-  color?: ListColor,
-  disabled?: boolean,
-  defaultColor?: string
-) => {
-  if (disabled) {
-    return 'text1'
-  } else if (color) {
-    if (listItemColorAppliesToLabel.includes(color)) {
-      // Theme "slot" & color is applied to label
-      return color
-    } else if (!listItemColorOptions.includes(color)) {
-      // HTML color
-      return color
-    }
-  }
-  return defaultColor
-}
-
-export const listItemIconColor = (color?: ListColor, disabled?: boolean) =>
-  listItemColor(color, disabled, 'text2')
-
-export const listItemLabelColor = (color?: ListColor, disabled?: boolean) =>
-  listItemColor(color, disabled, 'text5')
+export * from './getDetailOptions'
+export * from './listItemBackgroundColor'
+export * from './listItemColor'
+export * from './listItemDimensions'
+export * from './listItemPadding'
+export * from './createListItemPartitions'
