+++ packages/components/src/Panel/Panel.story.tsx
@@ -29,7 +29,8 @@ import { Story } from '@storybook/react/types-6-0'
 import { Done } from '@styled-icons/material/Done'
 import { Button } from '../Button'
 import { Drawer } from '../Drawer'
-import { List, ListItem } from '../List'
+import { List } from '../List'
+import { ListItem } from '../ListItem'
 import { Aside, Layout, Page, Section, SpaceVertical } from '../Layout'
 import { Paragraph } from '../Text'
 import { TreeCollection, Tree, TreeItem } from '../Tree'
