+++ packages/components/src/Tree/TreeStyle.tsx
@@ -29,15 +29,15 @@ import styled, { css } from 'styled-components'
 import { SpacingSizes } from '@looker/design-tokens'
 import { Accordion2Disclosure } from '../Accordion2/Accordion2Disclosure'
 import { Accordion2Content } from '../Accordion2/Accordion2Content'
-import { FlexibleColor, ListItemStatefulProps } from '../List/types'
-import { listItemBackgroundColor } from '../List/utils'
+import { FlexibleColor, ListItemStatefulProps } from '../ListItem/types'
+import { List } from '../List'
 import {
-  List,
   ListItem,
   ListItemContent,
   listItemContentCSS,
   ListItemIcon,
-} from '../List'
+} from '../ListItem'
+import { listItemBackgroundColor } from '../ListItem/utils'
 import { IconSize, IconType } from '../Icon'
 import { TreeAccordion } from './TreeAccordion'
 import { TreeItem } from './TreeItem'
