+++ packages/components/src/Tree/WindowedTreeCollection.tsx
@@ -33,7 +33,7 @@ import {
 } from '@looker/design-tokens'
 import React from 'react'
 import styled from 'styled-components'
-import { DensityProps } from '../List/types'
+import { DensityProps } from '../ListItem/types'
 import { useWindowedTree } from './utils/useWindowedTree'
 import { WindowedTreeNodeProps } from './types'
 import { TreeCollectionContext } from './TreeCollectionContext'
