+++ packages/components/src/Tree/stories/Density.story.tsx
@@ -28,7 +28,7 @@ import React, { FC } from 'react'
 import { Tree } from '../Tree'
 import { TreeCollection } from '../TreeCollection'
 import { TreeItem } from '../TreeItem'
-import { DensityRamp } from '../../List/types'
+import { DensityRamp } from '../../ListItem/types'
 import { Grid } from '../../Layout'
 
 const DensityTree: FC<{ density: DensityRamp }> = ({ density }) => (
