+++ packages/components/src/index.ts
@@ -44,6 +44,7 @@ export * from './Icon'
 export * from './Layout'
 export * from './Link'
 export * from './List'
+export * from './ListItem'
 export * from './Menu'
 export * from './MessageBar'
 export * from './NavList'
