+++ packages/components/src/utils/createSafeRel.ts
